agentic-samples/strands-observability-and-evaluation/strands_agent_bedrock_agentcore_with_observability/requirements.txt
# Core Strands and Agent Dependencies
strands-agents
strands-agents-tools

# Langfuse Observability
langfuse>=3.0.0

# AWS Dependencies
boto3
opensearch-py
botocore
bedrock-agentcore
bedrock-agentcore-starter-toolkit
aws-opentelemetry-distro==0.10.0